Removed dead code and renamed pipeline module
