+++ crates/gosub_engine/src/engine.rs
@@ -30,4 +30,4 @@ pub use policy::UaPolicy;
 /// Default capacity for MPSC channels
 const DEFAULT_CHANNEL_CAPACITY: usize = 512;
 
-pub mod pipeline;
+pub mod resource_pipeline;
