+++ crates/gosub_engine/src/engine/resource_pipeline.rs
@@ -1,12 +1,12 @@
-//! Pipeline modules for processing different asset types.
+//! Resource pipeline modules for processing different asset types.
 //!
 //! Each module defines a trait for parsing streams and byte slices of the respective asset type.
 
-use crate::engine::pipeline::css::{CssPipeline, CssPipelineImpl};
-use crate::engine::pipeline::font::{FontPipeline, FontPipelineImpl};
-use crate::engine::pipeline::html::{HtmlPipeline, HtmlPipelineImpl};
-use crate::engine::pipeline::image::{ImagePipeline, ImagePipelineImpl};
-use crate::engine::pipeline::js::{JsPipeline, JsPipelineImpl};
+use crate::engine::resource_pipeline::css::{CssPipeline, CssPipelineImpl};
+use crate::engine::resource_pipeline::font::{FontPipeline, FontPipelineImpl};
+use crate::engine::resource_pipeline::html::{HtmlPipeline, HtmlPipelineImpl};
+use crate::engine::resource_pipeline::image::{ImagePipeline, ImagePipelineImpl};
+use crate::engine::resource_pipeline::js::{JsPipeline, JsPipelineImpl};
 use crate::engine::types::IoChannel;
 use crate::zone::ZoneId;
 
@@ -16,9 +16,8 @@ pub mod html;
 pub mod image;
 pub mod js;
 
-/// Hooks are functions that allows the router to call the correct pipeline for each type of
-/// resource.
-pub struct Hooks {
+/// Resource pipeline entry points used by the router for each resource type.
+pub struct ResourcePipelines {
     pub html: Box<dyn HtmlPipeline + Send>,
     pub css: Box<dyn CssPipeline + Send>,
     pub js: Box<dyn JsPipeline + Send>,
@@ -29,7 +28,7 @@ pub struct Hooks {
     // pub external: &'a mut dyn ExternalOpener,
 }
 
-impl Hooks {
+impl ResourcePipelines {
     pub fn new(zone_id: ZoneId, io_tx: IoChannel) -> Self {
         Self {
             html: Box::new(HtmlPipelineImpl::new(zone_id, io_tx)),
