+++ crates/gosub_engine/src/engine/resource_pipeline/image.rs
