+++ crates/gosub_renderer_skia/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod compositor;
 pub(crate) mod font;
 pub mod rasterizer;
 
-pub use compositor::{skia_compose, SkiaCompositor};
 pub use rasterizer::SkiaRasterizer;
